pyproject.toml
[tool.poetry]
name = "biancabot"
version = "0.1.0"
description = ""
authors = ["Penloy <penloyone@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8.2"
pyyaml = "^5.3.1"
"discord.py" = "^1.3.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
